Delete more stale old pdata code
