+++ receiver/prometheusreceiver/internal/otlp_metricsbuilder.go
@@ -15,21 +15,124 @@
 package internal
 
 import (
+	"errors"
 	"fmt"
 	"regexp"
 	"sort"
 	"strconv"
+	"strings"
 
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/pkg/textparse"
 	"github.com/prometheus/prometheus/pkg/value"
-
 	"go.uber.org/zap"
 
 	"go.opentelemetry.io/collector/model/pdata"
 )
 
+const (
+	metricsSuffixCount  = "_count"
+	metricsSuffixBucket = "_bucket"
+	metricsSuffixSum    = "_sum"
+	metricSuffixTotal   = "_total"
+	startTimeMetricName = "process_start_time_seconds"
+	scrapeUpMetricName  = "up"
+)
+
+var (
+	trimmableSuffixes     = []string{metricsSuffixBucket, metricsSuffixCount, metricsSuffixSum, metricSuffixTotal}
+	errNoDataToBuild      = errors.New("there's no data to build")
+	errNoBoundaryLabel    = errors.New("given metricType has no BucketLabel or QuantileLabel")
+	errEmptyBoundaryLabel = errors.New("BucketLabel or QuantileLabel is empty")
+)
+
+// dpgSignature is used to create a key for data complexValue belong to a same group of a metric family
+func dpgSignature(orderedKnownLabelKeys []string, ls labels.Labels) string {
+	size := 0
+	for _, k := range orderedKnownLabelKeys {
+		v := ls.Get(k)
+		if v == "" {
+			continue
+		}
+		// 2 enclosing quotes + 1 equality sign = 3 extra chars.
+		// Note: if any character in the label value requires escaping,
+		// we'll need more space than that, which will lead to some
+		// extra allocation.
+		size += 3 + len(k) + len(v)
+	}
+	sign := make([]byte, 0, size)
+	for _, k := range orderedKnownLabelKeys {
+		v := ls.Get(k)
+		if v == "" {
+			continue
+		}
+		sign = strconv.AppendQuote(sign, k+"="+v)
+	}
+	return string(sign)
+}
+
+func normalizeMetricName(name string) string {
+	for _, s := range trimmableSuffixes {
+		if strings.HasSuffix(name, s) && name != s {
+			return strings.TrimSuffix(name, s)
+		}
+	}
+	return name
+}
+
+/*
+   code borrowed from the original promreceiver
+*/
+
+func heuristicalMetricAndKnownUnits(metricName, parsedUnit string) string {
+	if parsedUnit != "" {
+		return parsedUnit
+	}
+	lastUnderscoreIndex := strings.LastIndex(metricName, "_")
+	if lastUnderscoreIndex <= 0 || lastUnderscoreIndex >= len(metricName)-1 {
+		return ""
+	}
+
+	unit := ""
+
+	supposedUnit := metricName[lastUnderscoreIndex+1:]
+	switch strings.ToLower(supposedUnit) {
+	case "millisecond", "milliseconds", "ms":
+		unit = "ms"
+	case "second", "seconds", "s":
+		unit = "s"
+	case "microsecond", "microseconds", "us":
+		unit = "us"
+	case "nanosecond", "nanoseconds", "ns":
+		unit = "ns"
+	case "byte", "bytes", "by":
+		unit = "By"
+	case "bit", "bits":
+		unit = "Bi"
+	case "kilogram", "kilograms", "kg":
+		unit = "kg"
+	case "gram", "grams", "g":
+		unit = "g"
+	case "meter", "meters", "metre", "metres", "m":
+		unit = "m"
+	case "kilometer", "kilometers", "kilometre", "kilometres", "km":
+		unit = "km"
+	case "milimeter", "milimeters", "milimetre", "milimetres", "mm":
+		unit = "mm"
+	case "nanogram", "ng", "nanograms":
+		unit = "ng"
+	}
+
+	return unit
+}
+
+func isInternalMetric(metricName string) bool {
+	if metricName == scrapeUpMetricName || strings.HasPrefix(metricName, "scrape_") {
+		return true
+	}
+	return false
+}
 func isUsefulLabelPdata(mType pdata.MetricDataType, labelKey string) bool {
 	switch labelKey {
 	case model.MetricNameLabel, model.InstanceLabel, model.SchemeLabel, model.MetricsPathLabel, model.JobLabel:
@@ -84,9 +187,19 @@ func convToPdataMetricType(metricType textparse.MetricType) pdata.MetricDataType
 }
 
 type metricBuilderPdata struct {
-	*metricBuilder
-	metrics   pdata.MetricSlice
-	currentMf MetricFamilyPdata
+	metrics              pdata.MetricSlice
+	currentMf            MetricFamilyPdata
+	hasData              bool
+	hasInternalMetric    bool
+	mc                   MetadataCache
+	numTimeseries        int
+	droppedTimeseries    int
+	useStartTimeMetric   bool
+	startTimeMetricRegex *regexp.Regexp
+	startTime            float64
+	intervalStartTimeMs  int64
+	logger               *zap.Logger
+	stalenessStore       *stalenessStore
 }
 
 // newMetricBuilder creates a MetricBuilder which is allowed to feed all the datapoints from a single prometheus
@@ -98,17 +211,23 @@ func newMetricBuilderPdata(mc MetadataCache, useStartTimeMetric bool, startTimeM
 		regex, _ = regexp.Compile(startTimeMetricRegex)
 	}
 	return &metricBuilderPdata{
-		metrics: pdata.NewMetricSlice(),
-		metricBuilder: &metricBuilder{
-			mc:                   mc,
-			logger:               logger,
-			numTimeseries:        0,
-			droppedTimeseries:    0,
-			useStartTimeMetric:   useStartTimeMetric,
-			startTimeMetricRegex: regex,
-			stalenessStore:       stalenessStore,
-		},
+		metrics:              pdata.NewMetricSlice(),
+		mc:                   mc,
+		logger:               logger,
+		numTimeseries:        0,
+		droppedTimeseries:    0,
+		useStartTimeMetric:   useStartTimeMetric,
+		startTimeMetricRegex: regex,
+		stalenessStore:       stalenessStore,
 	}
+}
+
+func (b *metricBuilderPdata) matchStartTimeMetric(metricName string) bool {
+	if b.startTimeMetricRegex != nil {
+		return b.startTimeMetricRegex.MatchString(metricName)
+	}
+
+	return metricName == startTimeMetricName
 }
 
 // AddDataPoint is for feeding prometheus data complexValue in its processing order
@@ -171,9 +290,9 @@ func (b *metricBuilderPdata) AddDataPoint(ls labels.Labels, t int64, v float64) 
 		nTs, nDts := b.currentMf.ToMetricPdata(&b.metrics)
 		b.numTimeseries += nTs
 		b.droppedTimeseries += nDts
-		b.currentMf = newMetricFamilyPdata(metricName, b.mc, b.intervalStartTimeMs)
+		b.currentMf = newMetricFamilyPdata(metricName, b.mc, b.logger, b.intervalStartTimeMs)
 	} else if b.currentMf == nil {
-		b.currentMf = newMetricFamilyPdata(metricName, b.mc, b.intervalStartTimeMs)
+		b.currentMf = newMetricFamilyPdata(metricName, b.mc, b.logger, b.intervalStartTimeMs)
 	}
 
 	return b.currentMf.Add(metricName, ls, t, v)
