+++ receiver/prometheusreceiver/internal/otlp_metricsbuilder_test.go
@@ -15,97 +15,208 @@
 package internal
 
 import (
+	"runtime"
 	"testing"
 
-	metricspb "github.com/census-instrumentation/opencensus-proto/gen-go/metrics/v1"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/pkg/textparse"
+	"github.com/prometheus/prometheus/scrape"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/collector/model/pdata"
 )
 
-func TestGetBoundaryEquivalence(t *testing.T) {
-	cases := []struct {
-		name      string
-		mtype     metricspb.MetricDescriptor_Type
-		pmtype    pdata.MetricDataType
-		labels    labels.Labels
-		wantValue float64
-		wantErr   string
-	}{
+const startTs = int64(1555366610000)
+const interval = int64(15 * 1000)
+const defaultBuilderStartTime = float64(1.0)
+
+var testMetadata = map[string]scrape.MetricMetadata{
+	"counter_test":    {Metric: "counter_test", Type: textparse.MetricTypeCounter, Help: "", Unit: ""},
+	"counter_test2":   {Metric: "counter_test2", Type: textparse.MetricTypeCounter, Help: "", Unit: ""},
+	"gauge_test":      {Metric: "gauge_test", Type: textparse.MetricTypeGauge, Help: "", Unit: ""},
+	"gauge_test2":     {Metric: "gauge_test2", Type: textparse.MetricTypeGauge, Help: "", Unit: ""},
+	"hist_test":       {Metric: "hist_test", Type: textparse.MetricTypeHistogram, Help: "", Unit: ""},
+	"hist_test2":      {Metric: "hist_test2", Type: textparse.MetricTypeHistogram, Help: "", Unit: ""},
+	"ghist_test":      {Metric: "ghist_test", Type: textparse.MetricTypeGaugeHistogram, Help: "", Unit: ""},
+	"summary_test":    {Metric: "summary_test", Type: textparse.MetricTypeSummary, Help: "", Unit: ""},
+	"summary_test2":   {Metric: "summary_test2", Type: textparse.MetricTypeSummary, Help: "", Unit: ""},
+	"unknown_test":    {Metric: "unknown_test", Type: textparse.MetricTypeUnknown, Help: "", Unit: ""},
+	"poor_name_count": {Metric: "poor_name_count", Type: textparse.MetricTypeCounter, Help: "", Unit: ""},
+	"up":              {Metric: "up", Type: textparse.MetricTypeCounter, Help: "", Unit: ""},
+	"scrape_foo":      {Metric: "scrape_foo", Type: textparse.MetricTypeCounter, Help: "", Unit: ""},
+	"example_process_start_time_seconds": {Metric: "example_process_start_time_seconds",
+		Type: textparse.MetricTypeGauge, Help: "", Unit: ""},
+	"process_start_time_seconds": {Metric: "process_start_time_seconds",
+		Type: textparse.MetricTypeGauge, Help: "", Unit: ""},
+	"badprocess_start_time_seconds": {Metric: "badprocess_start_time_seconds",
+		Type: textparse.MetricTypeGauge, Help: "", Unit: ""},
+}
+
+type testDataPoint struct {
+	lb labels.Labels
+	t  int64
+	v  float64
+}
+
+type testScrapedPage struct {
+	pts []*testDataPoint
+}
+
+func createLabels(mFamily string, tagPairs ...string) labels.Labels {
+	lm := make(map[string]string)
+	lm[model.MetricNameLabel] = mFamily
+	if len(tagPairs)%2 != 0 {
+		panic("tag pairs is not even")
+	}
+
+	for i := 0; i < len(tagPairs); i += 2 {
+		lm[tagPairs[i]] = tagPairs[i+1]
+	}
+
+	return labels.FromMap(lm)
+}
+
+func createDataPoint(mname string, value float64, tagPairs ...string) *testDataPoint {
+	return &testDataPoint{
+		lb: createLabels(mname, tagPairs...),
+		v:  value,
+	}
+}
+
+func runBuilderStartTimeTests(t *testing.T, tests []buildTestDataPdata,
+	startTimeMetricRegex string, expectedBuilderStartTime float64) {
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			mc := newMockMetadataCache(testMetadata)
+			st := startTs
+			for _, page := range tt.inputs {
+				b := newMetricBuilderPdata(mc, true, startTimeMetricRegex, testLogger, dummyStalenessStore())
+				b.startTime = defaultBuilderStartTime // set to a non-zero value
+				for _, pt := range page.pts {
+					// set ts for testing
+					pt.t = st
+					assert.NoError(t, b.AddDataPoint(pt.lb, pt.t, pt.v))
+				}
+				_, _, _, err := b.Build()
+				assert.NoError(t, err)
+				assert.EqualValues(t, b.startTime, expectedBuilderStartTime)
+				st += interval
+			}
+		})
+	}
+}
+
+func Test_startTimeMetricMatch(t *testing.T) {
+	matchBuilderStartTime := 123.456
+	matchTests := []buildTestDataPdata{
 		{
-			name:   "cumulative histogram with bucket label",
-			mtype:  metricspb.MetricDescriptor_CUMULATIVE_DISTRIBUTION,
-			pmtype: pdata.MetricDataTypeHistogram,
-			labels: labels.Labels{
-				{Name: model.BucketLabel, Value: "0.256"},
+			name: "prefix_match",
+			inputs: []*testScrapedPage{
+				{
+					pts: []*testDataPoint{
+						createDataPoint("example_process_start_time_seconds",
+							matchBuilderStartTime, "foo", "bar"),
+					},
+				},
 			},
-			wantValue: 0.256,
 		},
 		{
-			name:   "gauge histogram with bucket label",
-			mtype:  metricspb.MetricDescriptor_GAUGE_DISTRIBUTION,
-			pmtype: pdata.MetricDataTypeHistogram,
-			labels: labels.Labels{
-				{Name: model.BucketLabel, Value: "11.71"},
+			name: "match",
+			inputs: []*testScrapedPage{
+				{
+					pts: []*testDataPoint{
+						createDataPoint("process_start_time_seconds",
+							matchBuilderStartTime, "foo", "bar"),
+					},
+				},
 			},
-			wantValue: 11.71,
+		},
+	}
+	nomatchTests := []buildTestDataPdata{
+		{
+			name: "nomatch1",
+			inputs: []*testScrapedPage{
+				{
+					pts: []*testDataPoint{
+						createDataPoint("_process_start_time_seconds",
+							matchBuilderStartTime, "foo", "bar"),
+					},
+				},
+			},
 		},
 		{
-			name:   "summary with bucket label",
-			mtype:  metricspb.MetricDescriptor_SUMMARY,
-			pmtype: pdata.MetricDataTypeSummary,
-			labels: labels.Labels{
-				{Name: model.BucketLabel, Value: "11.71"},
+			name: "nomatch2",
+			inputs: []*testScrapedPage{
+				{
+					pts: []*testDataPoint{
+						createDataPoint("subprocess_start_time_seconds",
+							matchBuilderStartTime, "foo", "bar"),
+					},
+				},
 			},
-			wantErr: "QuantileLabel is empty",
-		},
-		{
-			name:   "summary with quantile label",
-			mtype:  metricspb.MetricDescriptor_SUMMARY,
-			pmtype: pdata.MetricDataTypeSummary,
-			labels: labels.Labels{
-				{Name: model.QuantileLabel, Value: "92.88"},
-			},
-			wantValue: 92.88,
-		},
-		{
-			name:   "gauge histogram mismatched with bucket label",
-			mtype:  metricspb.MetricDescriptor_SUMMARY,
-			pmtype: pdata.MetricDataTypeSummary,
-			labels: labels.Labels{
-				{Name: model.BucketLabel, Value: "11.71"},
-			},
-			wantErr: "QuantileLabel is empty",
-		},
-		{
-			name:   "other data types without matches",
-			mtype:  metricspb.MetricDescriptor_GAUGE_DOUBLE,
-			pmtype: pdata.MetricDataTypeGauge,
-			labels: labels.Labels{
-				{Name: model.BucketLabel, Value: "11.71"},
-			},
-			wantErr: "given metricType has no BucketLabel or QuantileLabel",
 		},
 	}
 
-	for _, tt := range cases {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			oldBoundary, oerr := getBoundary(tt.mtype, tt.labels)
-			pdataBoundary, perr := getBoundaryPdata(tt.pmtype, tt.labels)
-			assert.Equal(t, oldBoundary, pdataBoundary, "Both boundary values MUST be equal")
-			assert.Equal(t, oldBoundary, tt.wantValue, "Mismatched boundary messages")
-			assert.Equal(t, oerr, perr, "The exact same error MUST be returned from both boundary helpers")
+	runBuilderStartTimeTests(t, matchTests, "^(.+_)*process_start_time_seconds$", matchBuilderStartTime)
+	runBuilderStartTimeTests(t, nomatchTests, "^(.+_)*process_start_time_seconds$", defaultBuilderStartTime)
+}
 
-			if tt.wantErr != "" {
-				require.NotEqual(t, oerr, "expected an error from old style boundary retrieval")
-				require.NotEqual(t, perr, "expected an error from new style boundary retrieval")
-				require.Contains(t, oerr.Error(), tt.wantErr)
-				require.Contains(t, perr.Error(), tt.wantErr)
+func Test_heuristicalMetricAndKnownUnits(t *testing.T) {
+	tests := []struct {
+		metricName string
+		parsedUnit string
+		want       string
+	}{
+		{"test", "ms", "ms"},
+		{"millisecond", "", ""},
+		{"test_millisecond", "", "ms"},
+		{"test_milliseconds", "", "ms"},
+		{"test_ms", "", "ms"},
+		{"test_second", "", "s"},
+		{"test_seconds", "", "s"},
+		{"test_s", "", "s"},
+		{"test_microsecond", "", "us"},
+		{"test_microseconds", "", "us"},
+		{"test_us", "", "us"},
+		{"test_nanosecond", "", "ns"},
+		{"test_nanoseconds", "", "ns"},
+		{"test_ns", "", "ns"},
+		{"test_byte", "", "By"},
+		{"test_bytes", "", "By"},
+		{"test_by", "", "By"},
+		{"test_bit", "", "Bi"},
+		{"test_bits", "", "Bi"},
+		{"test_kilogram", "", "kg"},
+		{"test_kilograms", "", "kg"},
+		{"test_kg", "", "kg"},
+		{"test_gram", "", "g"},
+		{"test_grams", "", "g"},
+		{"test_g", "", "g"},
+		{"test_nanogram", "", "ng"},
+		{"test_nanograms", "", "ng"},
+		{"test_ng", "", "ng"},
+		{"test_meter", "", "m"},
+		{"test_meters", "", "m"},
+		{"test_metre", "", "m"},
+		{"test_metres", "", "m"},
+		{"test_m", "", "m"},
+		{"test_kilometer", "", "km"},
+		{"test_kilometers", "", "km"},
+		{"test_kilometre", "", "km"},
+		{"test_kilometres", "", "km"},
+		{"test_km", "", "km"},
+		{"test_milimeter", "", "mm"},
+		{"test_milimeters", "", "mm"},
+		{"test_milimetre", "", "mm"},
+		{"test_milimetres", "", "mm"},
+		{"test_mm", "", "mm"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.metricName, func(t *testing.T) {
+			if got := heuristicalMetricAndKnownUnits(tt.metricName, tt.parsedUnit); got != tt.want {
+				t.Errorf("heuristicalMetricAndKnownUnits() = %v, want %v", got, tt.want)
 			}
 		})
 	}
@@ -1209,4 +1320,70 @@ func Test_metricBuilder_baddata(t *testing.T) {
 			t.Error("expecting errEmptyBoundaryLabel error, but get nil")
 		}
 	})
+}
+
+func Benchmark_dpgSignature(b *testing.B) {
+	knownLabelKeys := []string{"a", "b"}
+	labels := labels.FromStrings("a", "va", "b", "vb", "x", "xa")
+	b.ReportAllocs()
+	for i := 0; i < b.N; i++ {
+		runtime.KeepAlive(dpgSignature(knownLabelKeys, labels))
+	}
+}
+
+func Test_dpgSignature(t *testing.T) {
+	knownLabelKeys := []string{"a", "b"}
+
+	tests := []struct {
+		name string
+		ls   labels.Labels
+		want string
+	}{
+		{"1st label", labels.FromStrings("a", "va"), `"a=va"`},
+		{"2nd label", labels.FromStrings("b", "vb"), `"b=vb"`},
+		{"two labels", labels.FromStrings("a", "va", "b", "vb"), `"a=va""b=vb"`},
+		{"extra label", labels.FromStrings("a", "va", "b", "vb", "x", "xa"), `"a=va""b=vb"`},
+		{"different order", labels.FromStrings("b", "vb", "a", "va"), `"a=va""b=vb"`},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := dpgSignature(knownLabelKeys, tt.ls); got != tt.want {
+				t.Errorf("dpgSignature() = %q, want %q", got, tt.want)
+			}
+		})
+	}
+
+	// this is important for caching start values, as new metrics with new tag of a same group can come up in a 2nd run,
+	// however, its order within the group is not predictable. we need to have a way to generate a stable key even if
+	// the total number of keys changes in between different scrape runs
+	t.Run("knownLabelKeys updated", func(t *testing.T) {
+		ls := labels.FromStrings("a", "va")
+		want := dpgSignature(knownLabelKeys, ls)
+		got := dpgSignature(append(knownLabelKeys, "c"), ls)
+		if got != want {
+			t.Errorf("dpgSignature() = %v, want %v", got, want)
+		}
+	})
+}
+
+func Test_normalizeMetricName(t *testing.T) {
+	tests := []struct {
+		name  string
+		mname string
+		want  string
+	}{
+		{"normal", "normal", "normal"},
+		{"count", "foo_count", "foo"},
+		{"bucket", "foo_bucket", "foo"},
+		{"sum", "foo_sum", "foo"},
+		{"total", "foo_total", "foo"},
+		{"no_prefix", "_sum", "_sum"},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := normalizeMetricName(tt.mname); got != tt.want {
+				t.Errorf("normalizeMetricName() = %v, want %v", got, tt.want)
+			}
+		})
+	}
 }
